refactor(react-core): improve React ref handing in TemplateRenderer (#299)
