--- packages/dx-react-core/src/template-component.jsx
@@ -1,2 +1,0 @@
-export const TemplateRenderer = ({ template, children, ...restProps }) =>
-  template({ ...restProps, children });
