+++ packages/dx-react-core/src/template-renderer.jsx
@@ -1,0 +1,9 @@
+import PropTypes from 'prop-types';
+
+export const TemplateRenderer = ({ template, templateRef, ...restProps }) =>
+  template({ ...restProps, ref: templateRef });
+
+TemplateRenderer.propTypes = {
+  template: PropTypes.func.isRequired,
+  templateRef: PropTypes.func,
+};
