+++ packages/dx-react-core/src/template-renderer.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import { TemplateRenderer } from './template-component';
+import { TemplateRenderer } from './template-renderer';
 
 describe('TemplateRenderer', () => {
   it('should work', () => {
@@ -16,5 +16,19 @@ describe('TemplateRenderer', () => {
 
     expect(tree.find('.test > .content').exists())
       .toBe(true);
+  });
+
+  it('should pass templateRef as ref', () => {
+    const templateRef = jest.fn();
+    mount(
+      <TemplateRenderer
+        template={({ ref }) => <div ref={ref} />}
+        templateRef={templateRef}
+        test={'test'}
+      />,
+    );
+
+    expect(templateRef)
+      .toBeCalled();
   });
 });
